requirement.txt
arrow==1.3.0
boto3==1.19.12
filelock==3.12.4
lightning==2.1.2
lightning-cloud==0.5.52
lightning-utilities==0.10.0
markdown-it-py==3.0.0
pydantic==2.5.2
pydantic_core==2.14.5
pytorch-lightning==2.1.2
sentencepiece==0.1.99
wandb==0.15.3
zstandard==0.22.0
transformers==4.37.2
numpy==1.22.4
jsonargparse==4.32.0
backoff==2.2.1
beautifulsoup4==4.12.3
blessed==1.20.0
croniter==1.4.1
dateutils==0.6.12
deepdiff==6.7.1
editor==1.6.6
inquirer==3.4.0
itsdangerous==2.2.0
ordered-set==4.1.0
readchar==4.2.0
runs==1.2.2
soupsieve==2.6
starsessions==1.3.0
traitlets==5.14.3
wcwidth==0.2.13
websockets==11.0.3
xmod==1.8.1
datasets